// src/core/components/shared/dynamic-form/types/field-type.type.ts
/**
 * Defines the supported field types in the dynamic form system.
 * 
 * @type {FieldType}
 */
export type FieldType = 
  | 'text'
  | 'email'
  | 'password'
  | 'number'
  | 'tel'
  | 'url'
  | 'textarea'
  | 'richtext'
  | 'checkbox'
  | 'switch'
  | 'radio'
  | 'select'
  | 'date'
  | 'time'
  | 'datetime'
  | 'file'
  | 'array'
  | 'object'
  | 'address'
  | 'map'
  | 'voice'
  | 'custom';